simplify server.train_on_step; move AbstractClient into client
