+++ cezo_fl/client.py
@@ -1,16 +1,61 @@
+from __future__ import annotations
+
 import torch
 from torch.utils.data import DataLoader
 from typing import Sequence
 from copy import deepcopy
+import abc
+from dataclasses import dataclass
 
 from shared.metrics import Metric
 from gradient_estimators.random_gradient_estimator import RandomGradientEstimator as RGE
-from cezo_fl.server import AbstractClient, LocalUpdateResult
 from cezo_fl.shared import (
     CriterionType,
     update_model_given_seed_and_grad,
     revert_SGD_given_seed_and_grad,
 )
+
+
+@dataclass
+class LocalUpdateResult:
+    grad_tensors: list[torch.Tensor]
+    step_accuracy: float
+    step_loss: float
+
+    # Must add __future__ import to be able to return, see https://stackoverflow.com/a/33533514
+    def to(self, device: torch.device) -> LocalUpdateResult:
+        self.grad_tensors = [grad_tensor.to(device) for grad_tensor in self.grad_tensors]
+        return self
+
+
+class AbstractClient:
+
+    @abc.abstractmethod
+    def local_update(self, seeds: Sequence[int]) -> LocalUpdateResult:
+        """Returns a sequence of gradient scalar tensors for each local update.
+
+        The length of the returned sequence should be the same as the length of seeds.
+        The inner tensor can be a scalar or a vector. The length of vector is the number
+        of perturbations.
+        """
+        return NotImplemented
+
+    @abc.abstractmethod
+    def reset_model(self) -> None:
+        """Reset the mode to the state before the local_update."""
+        return NotImplemented
+
+    @abc.abstractmethod
+    def pull_model(
+        self,
+        seeds_list: Sequence[Sequence[int]],
+        gradient_scalar: Sequence[Sequence[torch.Tensor]],
+    ) -> None:
+        return NotImplemented
+
+    @abc.abstractmethod
+    def random_gradient_estimator(self) -> RGE:
+        return NotImplemented
 
 
 class SyncClient(AbstractClient):
@@ -40,14 +85,14 @@ class SyncClient(AbstractClient):
         self.local_update_seeds: list[int] = []
         self.local_update_dir_grads: list[torch.Tensor] = []
 
-    def random_gradient_estimator(self):
-        return self.grad_estimator
-
     def _get_train_batch_iterator(self):
         # NOTE: used only in init, will generate an infinite iterator from dataloader
         while True:
             for v in self.dataloader:
                 yield v
+
+    def random_gradient_estimator(self):
+        return self.grad_estimator
 
     def local_update(self, seeds: Sequence[int]) -> LocalUpdateResult:
         """Returns a sequence of gradient scalar tensors for each local update.
